исправлено обновление данных
