+++ src/data.js
@@ -139,20 +139,9 @@ const getEvent = () => {
 
 export const getEventsData = (count) => {
   const events = new Array(count);
-  return events.fill(``).map(getEvent).sort((a, b) => a.start - b.start);
+  return events.fill(``).map(getEvent);
 };
 
-export const getEventsInDays = (eventsData) => {
-  return eventsData.reduce((acc, event) => {
-    const date = new Date(event.start).toDateString();
-    if (acc[date]) {
-      acc[date].push(event);
-    } else {
-      acc[date] = [event];
-    }
-    return acc;
-  }, {});
-};
 
 export const getCities = (eventsData) => {
   return eventsData.map((event) => event.city);
