+++ src/util.js
@@ -55,3 +55,15 @@ export const formatDate = (date) => {
   let year = date.getFullYear() % 100;
   return `${day < 10 ? `0${day}` : day}.${month < 10 ? `0${month}` : month}.${year < 10 ? `0${year}` : year}`;
 };
+export const getEventsInDays = (eventsData) => {
+  return eventsData.reduce((acc, event) => {
+    const date = new Date(event.start).toDateString();
+    if (acc[date]) {
+      acc[date].push(event);
+    } else {
+      acc[date] = [event];
+    }
+    return acc;
+  }, {});
+};
+
